+++ test/codegen/end_to_end_test.dart
@@ -121,6 +121,10 @@ void main(List<String> args) {
     runQuery('K', new QueryOptions(
         includeInherited: true, withAnnotations: [lib.getType('AnnotC')]));
 
+    runQuery('J', new QueryOptions(includeMethods: true));
+    runQuery('J2', new QueryOptions(
+        includeInherited: true, includeMethods: true));
+
     var code = _createEntrypoint(generator);
     var staticTestFile = new File(path.join(testDir, 'static_test.dart'));
     var existingCode = staticTestFile.readAsStringSync();