+++ test/mirrors_used_test.dart
@@ -10,7 +10,7 @@ import 'common.dart' hide main;
 import 'common.dart' as common show main;
 
 @MirrorsUsed(
-    targets: const [A, B, C, D, E, E2, F, F2, G, H, Type],
+    targets: const [A, B, C, D, E, E2, F, F2, G, H, J, J2, Type],
     override: 'smoke.mirrors')
 import 'dart:mirrors';