+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#### 0.3.4
+  * Add excludeOverriden to QueryOptions which removes declarations that were
+    overriden by a base class.
+
 #### 0.3.3+1
   * Update logging package to `<0.12.0`.